+++ crates/affinidi-tdk/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Affinidi TDK Changelog
 
+## 24th March 2025 (release 0.1.6)
+
+* Added add_profile() to TDK and TDK-Common. Allows for loading of secrets
+
 ## Release 0.1.0
 
 * Initial release of crate